% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leading_char.R
\name{add_leading_char}
\alias{add_leading_char}
\title{Add leading zeroes to strings}
\usage{
add_leading_char(target, n, char = "0")
}
\arguments{
\item{target}{string or vector of strings to add leading zero (zeros) to.}

\item{n}{number of characters the returned strings should be.}
}
\value{
This function returns a string or vector of strings with leading zeros
}
\description{
Add leading zeroes to strings
}
\author{
Bjorn Kallerud
}
