% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_consumption.R
\name{allocate_consumption}
\alias{allocate_consumption}
\title{Allocate Tiered Water Consumption}
\usage{
allocate_consumption(df, suffix = NULL, use.prime = FALSE)
}
\arguments{
\item{df}{dataframe that contains customer usage and tier widths. usage column can be titled either "use" or "usage". width columns titles must take the form of "tX_width".}

\item{suffix}{width column suffix. we often define widths in terms of current/proposed rates (ie tX_width_current), so defining the suffix enables use of this function for current and proposed widths.}

\item{use.prime}{boolean - set to TRUE if you want to use `use_prime` usage column instead of standard use column.}
}
\value{
This function returns a \code{dataframe} of numeric columns for tiered usage
}
\description{
Distributes a customers total water consumption over various tiers.
}
\author{
Bjorn Kallerud
}
