% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bills.R
\name{compare_bills}
\alias{compare_bills}
\title{Compare Water Bill}
\usage{
compare_bills(df_use, df_rates, rate_group = c("class", "meter_size"),
  use.essential = 2)
}
\arguments{
\item{df_use}{\code{dataframe} that contains customer usage - var must be titled \code{usage} or \code{use}.}

\item{df_rates}{\code{dataframe} that contains current and proposed fixed and volumetric rates.}

\item{rate_group}{\code{vector} of strings for the group that defines unique rate structure.}

\item{use.essential}{SFR usage (in same units as use) that is deemed essential. All usage for classes other than SFR is categorized as essential.}
}
\value{
This function returns a \code{dataframe} that includes all final bills under the current and proposed rate
structures as well as their respective tier components.
}
\description{
Calculates customers' water bills based on fixed and volumetric rates. There is an option to include both current
and proposed rate structures to assess bill impacts.
}
\details{
BillCalculation <- function(df, elasticity = 0.0, nt_c, nt_p = NULL, uniform_rates = TRUE, proposed_rates = TRUE,
}
\author{
Bjorn Kallerud
}
