% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_meters.R
\name{convert_meters}
\alias{convert_meters}
\title{Convert Meter Sizes}
\usage{
convert_meters(meter_col, fraction = TRUE)
}
\arguments{
\item{meter_col}{vector of meter sizes to convert}

\item{fraction}{if TRUE, convert meter sizes to fraction format}
}
\value{
This function returns a vector of correctly formatted meter sizes
}
\description{
Standardizes meter size formatting
}
\author{
Zach Van Dinther
}
