% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert Water Units}
\usage{
convert_units(xx, from = "gallons", to = "hcf")
}
\arguments{
\item{xx}{numeric value or vector to be converted}

\item{from}{unit to convert from, either "gallons", "hcf", or "af" (acre-feet)}

\item{to}{unit to convert to, either "gallons", "hcf", or "af" (acre-feet)}
}
\value{
This function returns a numeric value or vector
}
\description{
Converts between gallons, hcf, and acre-feet
}
\author{
Bjorn Kallerud
}
