% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastMeters.R
\name{ForecastMeters}
\alias{ForecastMeters}
\title{Forecast Meters}
\usage{
ForecastMeters(df, t_0 = 2018, t_final = 2024)
}
\arguments{
\item{df}{dataframe that contains annual observations of the number of meters by size}

\item{t_0}{integer for final year for which data is available}

\item{t_final}{integer for the final year of forecasting period}
}
\value{
This function returns a \code{dataframe} of forecasted meters by group
}
\description{
\code{dplyr}-style function that forecasts the number of meters by group (typically meter size)
}
\author{
Bjorn Kallerud
}
