% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastUsage.R
\name{ForecastUsage}
\alias{ForecastUsage}
\title{Forecast Meters}
\usage{
ForecastUsage(df, t_0 = 2018, t_final = 2024)
}
\arguments{
\item{df}{dataframe that contains annual observations of per-meter usage}

\item{t_0}{integer for final year for which data is available}

\item{t_final}{integer for the final year of forecasting period}
}
\value{
This function returns a \code{dataframe} of forecasted per-meter usage
}
\description{
\code{dplyr}-style function that forecasts the number of meters by group (typically meter size)
}
\author{
Bjorn Kallerud
}
