% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TierConsumption.R
\name{TierConsumption}
\alias{TierConsumption}
\title{Calculate Tier Use}
\usage{
TierConsumption(nt, usage, t1width, t2width = NULL, t3width = NULL,
  t4width = NULL, t5width = NULL)
}
\arguments{
\item{nt}{integer that contains the number of tiers of the rate structure}

\item{usage}{numeric vector containing customer's water usage}

\item{widths}{integer that contains the width of each tier or a numeric vector that contains individual tier widths in the case of water budget rates}
}
\value{
This function returns a \code{list} of numeric vectors of water usage within each tier
}
\description{
Distributes a customers total water consumption over various tiers
}
\author{
Bjorn Kallerud
}
