% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBill.R
\name{WaterBill}
\alias{WaterBill}
\title{Water Bill Calculation}
\usage{
WaterBill(
  df,
  uniform_rates = TRUE,
  nt_c,
  wdths_c = NULL,
  fxd_c = NULL,
  vol_c = NULL,
  proposed_rates = TRUE,
  elasticity_par = 0,
  elasticity_response = "average",
  nt_p = NULL,
  wdths_p = NULL,
  fxd_p = NULL,
  vol_p = NULL
)
}
\arguments{
\item{df}{\code{dataframe} that contains customer usage - var must be titled \code{usage}. If rates are not uniform across
customers, fixed rate columns must be titled \code{fixed_c} and \code{fixed_p}}

\item{uniform_rates}{\code{logical} if \code{TRUE} the rates are constant across all customers in dataframe. If \code{TRUE}
inputs for the rate structure are required in the function call}

\item{nt_c}{number of tiers in the current rate structure}

\item{wdths_c}{\code{vector} of tier widths under the current rate structure - not necessary if rate structure is not uniform}

\item{fxd_c}{\code{double} of fixed charge under current rate structure - not necessary if rate structure is not uniform}

\item{vol_c}{\code{vector} of volumetric rates under the current rate structure - not necessary if rate structure is not uniform}

\item{proposed_rates}{\code{logical} if \code{TRUE} two rate structures are included in the calculation}

\item{elasticity_par}{the elasticity parameter, default is set to zero which excludes elasticity from the analysis}

\item{nt_p}{number of tiers in the proposed rate structure}

\item{wdths_p}{\code{vector} of tier widths under the proposed rate structure - not necessary if rate structure is not uniform}

\item{fxd_p}{\code{double} of fixed charge under proposed rate structure - not necessary if rate structure is not uniform}

\item{vol_p}{\code{vector} of volumetric rates under the proposed rate structure - not necessary if rate structure is not uniform}
}
\value{
This function returns a \code{dataframe} that includes all final bills under the current and proposed rate
structures as well as their respective components
}
\description{
Calculates customers' water bills based on fixed and volumetric rates. There is an option to include both current
and proposed rate structures to assess bill impacts.
}
\details{
BillCalculation <- function(df, elasticity = 0.0, nt_c, nt_p = NULL, uniform_rates = TRUE, proposed_rates = TRUE,
}
\author{
Bjorn Kallerud
}
