% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterr.R
\docType{package}
\name{waterr}
\alias{waterr}
\alias{waterr-package}
\title{waterr: Common Manipulations of Water Bill Data}
\description{
Perform common manipulations on customer-level water bill data
}
\author{
Bjorn Kallerud \email{bjorn@rdniehaus.com}
}
