% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bill.R
\name{calculate_bill}
\alias{calculate_bill}
\title{Calculate Water Bill}
\usage{
calculate_bill(
  df_use,
  df_rates,
  rate_group = c("class", "meter_size"),
  suffix = "current"
)
}
\arguments{
\item{df_use}{\code{dataframe} that contains customer usage - var must be titled \code{usage} or \code{use}.}

\item{df_rates}{\code{dataframe} that contains fixed and volumetric rates with suffix "_current". Set to \code{NULL} if widths and rates are already in \code{df_use}.}

\item{rate_group}{\code{vector} of strings for the group that defines unique rate structure.}

\item{suffix}{calculate bill based on \code{current} or \code{proposed} rate structure.}
}
\value{
This function returns a \code{dataframe} that includes the final bills under the current rate
structure as well as the bill's respective tier components.
}
\description{
Calculates customers' water bills based on fixed and volumetric rates. There is an option to include both current
and proposed rate structures to assess bill impacts.
}
\details{
calculate_bill <- function(df_use, df_rates, rate_group = c("class", "meter_size"), suffix = "current")
}
\author{
Bjorn Kallerud
}
